requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
packaging==25.0
pip-review==1.3.0
protobuf==6.33.5
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
openpyxl==3.1.5
